Halting development
Things are working sort-of right now. You can open a web page on the
display wall, which will use 28 projectors to display it. By running the
``inputman.go`` script on the different tiles, it is possible to recieve
user input from the /controller page.

However... This approach will not work because the webpage the user wants to
display is loaded into an iframe. Doing so it is impossible to stop the
tiles from messing up their location etc. From my best guess, it will
probably be easier to write a user script and using tampermonkey or
something, than this approach. By writing a user script you can inject code
directly into a web page, and there is no iframe mess to deal with.

If I get the time I will definately continue development and check out the
possibility of writing a user script